Rename(리뷰 내용 반영): 리뷰 내용 반영에 따른 함수 관리 파일 이름 변경
